move components files to folder screen
