+++ src/screens/Home/components/Partners/CardsPartnersHome/card.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import Typography from '@mui/material/Typography';
-import { Partner } from '../../interfaces/Partner';
+import { Partner } from '../../../../../lib/interfaces/Partner';
 import './styles.css';
 
 interface Props {
