+++ src/screens/Home/components/Partners/CardsPartnersHome/styles.css
