+++ src/screens/Home/components/Partners/PartnersCarousel/styles.css
