+++ src/screens/Home/index.tsx
@@ -5,12 +5,12 @@ import MyCarousel from '../../lib/components/Carousel/carousel.tsx';
 import { Container, Typography } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import SearchBar from '../../lib/components/SearchBar/searchbar.tsx';
-import PartnersCarousel from '../../lib/components/PartnersCarousel/partnerscarousel.tsx';
+import PartnersCarousel from './components/Partners/PartnersCarousel/partnerscarousel.tsx';
 import { getHomePagePartners } from '../../services/partners/index.ts';
 import { getHomePageTestimonials } from '../../services/testimonials/index.ts';
 import { getAllProductsCategories } from '../../services/categories/index.ts';
 import { Partner } from '../../lib/interfaces/Partner.ts';
-import TestimonialsCarousel from '../../lib/components/Testimonials/TestimonialsCarousel/carousel.tsx';
+import TestimonialsCarousel from './components/Testimonials/TestimonialsCarousel/carousel.tsx';
 import { Testimonial } from '../../lib/interfaces/Testimonial.ts';
 import { ProductCategory } from '../../lib/interfaces/Categories.ts';
 import SmallFooter from '../../lib/components/Footer/smallFooter.tsx';
